Add dates to receipts generated by standard sample data script
Previously sample receipts were created with the time being datetime.now. This made testing of the amount of data used to fetch receipts difficult as fetching receipts after a certain date was redundant because all receipts were on the same date.
This change fixes this by incrementing the date assigned to sample receipts every second receipt